feat: add full CRUD views (Baies, Offres, Clients, Réservations) + fixes
- Authentification : BCrypt → Argon2id (compatible Symfony security.yaml)
- SceneManager : setRoot() au lieu de setScene() → plus de flash au changement de page
- Baies : CRUD complet, modification du nombre d'unités, PieChart occupation
- Offres : CRUD complet, auto-calc prix annuel = mensuel × 10, format Double
- Clients : fix LazyInitializationException (@Transactional + init collections)
- Réservations : validation/annulation paiements, mise à jour Order + Payment
- Prix en centimes : affichage ÷ 100 (180000 → 1 800,00 €)
- Toutes les TableView : CONSTRAINED_RESIZE_POLICY (plus de barre blanche)
- ConfirmDialog : boîte de confirmation custom thème sombre

+++ src/main/java/org/example/wtg/entities/Offre.java
@@ -14,10 +14,10 @@ public class Offre {
     private Integer nombreUnites;
 
     @Column(name = "prix_mensuel")
-    private Integer prixMensuel;
+    private Double prixMensuel;
 
     @Column(name = "prix_annuel")
-    private Integer prixAnnuel;
+    private Double prixAnnuel;
 
     public Integer getId() {
         return id;
@@ -43,19 +43,19 @@ public class Offre {
         this.nombreUnites = nombreUnites;
     }
 
-    public Integer getPrixMensuel() {
+    public Double getPrixMensuel() {
         return prixMensuel;
     }
 
-    public void setPrixMensuel(Integer prixMensuel) {
+    public void setPrixMensuel(Double prixMensuel) {
         this.prixMensuel = prixMensuel;
     }
 
-    public Integer getPrixAnnuel() {
+    public Double getPrixAnnuel() {
         return prixAnnuel;
     }
 
-    public void setPrixAnnuel(Integer prixAnnuel) {
+    public void setPrixAnnuel(Double prixAnnuel) {
         this.prixAnnuel = prixAnnuel;
     }
 }
