+++ src/main/java/org/example/wtg/entities/Order.java
@@ -13,7 +13,7 @@ public class Order {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
 
-    private Integer total;
+    private Double total;
     private String status;
 
     @Column(name = "created_at")
@@ -38,11 +38,11 @@ public class Order {
         this.id = id;
     }
 
-    public Integer getTotal() {
+    public Double getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Double total) {
         this.total = total;
     }
